backend/pyproject.toml
[project]
name = "iit"
version = "0.1.0"
description = "Parallel agent orchestration system for data quality analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.2.0",
]
